Change pthread_cleanup_{push,pop} to macros that store the cleanup info on the stack instead of mallocing the list and move the APIs from libpthread to libc so that they can be used inside libc.
Note: the standard was explicitly written to permit/support this
"macro with unmatched brace" style and it's what basically everyone
else already does.  We xor the info with random cookies with a
random magic to detect/trip-up overwrites.

Major bump to both libc and libpthread due to the API move.

ok mpi@